% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa2r.R
\name{kappa2r}
\alias{kappa2r}
\title{Calculate Cohen's kappa with the confidence interval}
\usage{
kappa2r(v1, v2, weight = "equal", nboot = 500, cluster = FALSE, clu.id = NULL)
}
\arguments{
\item{v1, v2}{vectors, categorical ratings}

\item{weight}{one of c("equal","squared", "unweighted"), default is "equal"}

\item{cluster}{TRUE/FALSE, whether the data is clustered}

\item{clu.id}{subject id, when data is clustered}
}
\value{
a vector for the estimated kappa with the 95\%CI
}
\description{
The function estimates Cohen's kappa for 2 ratings with bootstrapped confidence interval.
}
\details{
kappa2r calculates Cohen's kappa with 95\% bootstrapped confidence intervals.
The confidence interval is the bias corrected ("bca") confidence interval. For clustered data, use kappa2r.clu
Check the categories in 2 ratings before running the function.
You should have them categoried in the same order.
}
\examples{
set.seed(234); x1 <- rbinom(n=50, size=2, prob=0.3)
set.seed(393); x2 <- rbinom(n=50, size=2, prob=0.4)
id <- c(1:25, 1:25)
kappa2r(x1,x2, weight="squared", nboot=500, cluster=TRUE, clu.id = id)

}
