% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkappamr.R
\name{lkappamr}
\alias{lkappamr}
\title{Calculate Lights's kappa with the confidence interval}
\usage{
lkappamr(ratings, weight = c("equal", "unweighted", "squared"))
}
\arguments{
\item{ratings}{a matrix, categorical ratings}

\item{weight}{one of c("equal","squared", "unweighted"), default is "equal"}
}
\value{
a list including n, the number of ratings, Light's kappa, and a p value
}
\description{
The function estimates Light's kappa for multiple ratings with the confidence interval.
Weights can be specified.
}
\details{
kappa2r calculates Cohen's kappa with 95\% bootstrapped confidence intervals.
The confidence interval is the bias corrected ("bca") confidence interval. For clustered data, use kappa2r.clu
Check the categories in every rating are in the same order.
}
\examples{
set.seed(234); x1 <- rbinom(n=50, size=2, prob=0.3)
set.seed(393); x2 <- rbinom(n=50, size=2, prob=0.4)
set.seed(395); x3 <- rbinom(n=50, size=2, prob=0.2)
lkappamr(cbind(x1,x2,x3), weight="equal")

}
