% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbaccu.lr.R
\name{tbaccu.lr}
\alias{tbaccu.lr}
\title{Table the measures of accuracy}
\usage{
tbaccu.lr(var, reference, thres = NULL, ci = TRUE)
}
\arguments{
\item{var}{a vector, test variable}

\item{reference}{a binary reference variable}

\item{thres}{a numeric value, the threshold value when the test variable is continuous}

\item{ci}{TRUE/FALSE, whether to include the confidence interval in the output}
}
\value{
a data frame for the estimates, with the 95\%CI (optional), and the counts
}
\description{
The function estimates the measures of accuracy (without adjustment)
for a binary reference variable and a test variable.
The test variable, when applicable, are dichotomized at the threshold value.
}
\details{
tbaccu.sslr outputs the table including estimates of sensitivity, specificity, LR+ and LR-.
Clustered data is not accepted in this function yet.
}
\examples{
y <- rbinom(n=50, size=1, prob=0.3) # reference variable
x <- rnorm(n=50, mean=20, sd=5)
tbaccu.lr(x,y, thres = 20)

}
