% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbaccu.R
\name{tbaccu.sspv}
\alias{tbaccu.sspv}
\title{Table the measures of accuracy}
\usage{
tbaccu.sspv(
  var,
  reference,
  thres = NULL,
  ci = TRUE,
  cluster = FALSE,
  id = NULL
)
}
\arguments{
\item{var}{a vector, test variable}

\item{reference}{a binary reference variable}

\item{thres}{a numeric value, the threshold value when the test variable is continuous}

\item{ci}{TRUE/FALSE, whether to include the confidence interval in the output}

\item{cluster}{TRUE/FALSE, an indicator of whether the data is clustered. The data was assume to adopt simple random sampling with equal weights}

\item{id}{subject ID variable, when data is clustered}
}
\value{
a data frame for the estimates, with the 95\%CI (optional), and the counts
}
\description{
The functions estimate the measures of accuracy (without adjustment)
for a binary reference variable and a test variable.
The test variable, when applicable, are dichotomized at the threshold value.
When the data is clustered, the survey package is used to estimate the confidence interval.
}
\details{
tbaccu.sspv outputs the table including estimates of sensitivity, specificity, PPV and NPV.
tbaccu.sslr outputs the table including estimates of sensitivity, specificity, LR+ and LR-.
Clustered data is not accepted in this function yet.
}
\examples{
y <- rbinom(n=50, size=1, prob=0.3) # reference variable
x <- rnorm(n=50, mean=20, sd=5)
tbaccu.sspv(x,y, thres = 20)
tbaccu.sslr(x,y, thres = 20)

}
