% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbyouden.R
\name{youdentb}
\alias{youdentb}
\title{Table the AUC, threshold, and measures of accuracy at the threshold}
\usage{
youdentb(var, reference)
}
\arguments{
\item{var}{a continuous predictor}

\item{reference}{a binary reference variable}
}
\value{
a data frame for the estimate, and the 95\%CI, the number when applicable
}
\description{
This function estimates the AUC for a binary reference variable and a continuous predictor,
reporting the threshold value maximizing the Youden's index,
and the corresponding measures of accuracy (without adjustment).
}
\examples{
y <- rbinom(n=50, size=1, prob=0.3) # reference variable
x <- rnorm(n=50, mean=20, sd=5)
youdentb(x,y)

}
