% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youdentb.R
\name{youdentb}
\alias{youdentb}
\title{Table the AUC, threshold, and measures of accuracy at the threshold}
\usage{
youdentb(var, reference)
}
\arguments{
\item{var}{a continuous predictor}

\item{reference}{a binary reference variable}
}
\value{
a vector
}
\description{
This function estimate the AUC for a binary reference variable and a continuous predictor,
reporting the threshold value maximizing the Youden's index,
and the corresponding measures of accuracy (without adjustment).
}
